#35 - TradeGenerator 유닛 테스트코드 작성
TradeGenerator 유닛 테스트코드 작성
- TransactionGenerator 객체에 들어갈 메서드 단위 테스트 코드 작성

+++ generator-core/src/main/java/com/simpaylog/generatorcore/cache/dto/TradeInfo.java
@@ -2,7 +2,6 @@ package com.simpaylog.generatorcore.cache.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record TradeInfo(
@@ -12,7 +11,7 @@ public record TradeInfo(
     // category 배열 내의 각 카테고리 정보를 위한 중첩 레코드
     public record CategoryDetail(
             @JsonProperty("name") String name,
-            @JsonProperty("weights") List<BigDecimal> weights,
+            @JsonProperty("weights") List<Double> weights,
             @JsonProperty("trades") List<TradeItemDetail> trades
     ) {}
